requirements.txt
pyside6
Pillow
pyinstaller
